Move config-loading out of helpers.
This change is due to that module being more important now, and not just
a helper.
